requirements.txt
paramiko
ftblib
pysmb
nmap
scapy
subprocess